@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:provider/single_child_widget.dart';
 import 'package:servplatform/core/enums/connectivity_status.dart';
 import 'package:servplatform/core/services/connectivity/connectivity_service.dart';
 import 'package:servplatform/locator.dart';
@@ -7,17 +8,17 @@ import 'package:servplatform/locator.dart';
 /// List of providers that provider transforms into a widget tree
 /// with the main app as the child of that tree, so that the entire
 /// app can use these streams anywhere there is a [BuildContext]
-List<SingleChildCloneableWidget> providers = [
+List<SingleChildWidget> providers = [
   ...independentServices,
   ...dependentServices,
   ...uiConsumableProviders
 ];
 
-List<SingleChildCloneableWidget> independentServices = [];
+List<SingleChildWidget> independentServices = [];
 
-List<SingleChildCloneableWidget> dependentServices = [];
+List<SingleChildWidget> dependentServices = [];
 
-List<SingleChildCloneableWidget> uiConsumableProviders = [
+List<SingleChildWidget> uiConsumableProviders = [
   StreamProvider<ConnectivityStatus>(
     create: (context) => locator<ConnectivityService>().connectivity$,
   ),
