@@ -5,12 +5,13 @@ import 'package:provider_architecture/provider_architecture.dart';
 import 'package:servplatform/core/enums/view_state.dart';
 import 'package:servplatform/core/localization/localization.dart';
 import 'package:servplatform/core/view_models/home_view_model.dart';
+import 'package:servplatform/core/view_models/merchant_home_view_model.dart';
 import 'package:servplatform/ui/widgets/list_header.dart';
 import 'package:servplatform/ui/widgets/search_field.dart';
 import 'package:servplatform/ui/widgets/service_tile.dart';
 import 'package:servplatform/ui/widgets/sliver_multiline_app_bar.dart';
 
-class HomeView extends StatelessWidget {
+class MerchantView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final local = AppLocalizations.of(context);
@@ -18,41 +19,50 @@ class HomeView extends StatelessWidget {
     final bottomAppBarTheme = Theme.of(context).bottomAppBarTheme;
     final iconTheme = Theme.of(context).iconTheme;
 
-    return ViewModelProvider<HomeViewModel>.withoutConsumer(
-        viewModel: HomeViewModel(),
+    return ViewModelProvider<MerchantHomeViewModel>.withoutConsumer(
+        viewModel: MerchantView(),
         onModelReady: (model) => model.init(),
         builder: (context, model, child) => Scaffold(
-              body: CustomScrollView(slivers: <Widget>[
-                SliverMultilineAppBar(
-                  title: 'Choose a Service for you below',
-                  subtitle: 'or swipe up to browse 2, 124 active services',
-                ),
-                SliverList(
-                  delegate: SliverChildListDelegate([SearchField()]),
-                ),
-                _RecommendedServices(),
-              ]),
-              bottomNavigationBar: BottomAppBar(
-                elevation: bottomAppBarTheme.elevation,
-                child: Row(
-                  mainAxisSize: MainAxisSize.max,
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  children: <Widget>[
-                    Text('More', style: textTheme.caption),
-                    LiveIconButton(
-                      icon: AnimatedIcons.menu_close,
-                      onPressed: () {},
-                    ),
-                  ],
-                ),
-              ),
-            ));
-  }
-}
+          body: CustomScrollView(slivers: <Widget>[
+            SliverMultilineAppBar(
+              title: 'Merchant',
+            ),
+            SliverList(
+              delegate: SliverChildListDelegate([SearchField()]),
+            ),
+            OptionalServices(),
+          ],
+            GroupedBarChart(),
+          ),
 
-class _RecommendedServices extends ProviderWidget<HomeViewModel> {
+          bottomNavigationBar: BottomAppBar(
+            elevation: bottomAppBarTheme.elevation,
+            child: Row(
+              mainAxisSize: MainAxisSize.max,
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: <Widget>[
+                Text('More', style: textTheme.caption),
+                LiveIconButton(
+                  icon: AnimatedIcons.menu_close,
+                  onPressed: () {},
+                ),
+              ],
+    floatingActionButton: FloatingActionButton(
+    onPressed: () {
+    // Set up more agencies
+    },
+    children: Icon(Icons.add),
+    Text('Set up Agencies',style: textTheme.caption),
+    backgroundColor: Colors.blueAccent,
+            ),
+          ),
+        ));
+  }
+
+
+class OptionalServices extends ProviderWidget<MerchantHomeViewModel> {
   @override
-  Widget build(BuildContext context, HomeViewModel model) {
+  Widget build(BuildContext context,MerchantViewModel model) {
     if (model.state == ViewState.Busy) {
       return _LoadingAnimation();
     }
@@ -63,7 +73,7 @@ class _RecommendedServices extends ProviderWidget<HomeViewModel> {
 
     return SliverList(
       delegate: SliverChildBuilderDelegate(
-        (context, index) {
+            (context, index) {
           return ServiceTile(
             key: Key('${model.services[index].id}'),
             service: model.services[index],
