@@ -15,48 +15,46 @@ class SearchField extends StatelessWidget {
   Widget build(BuildContext context) {
     final textTheme = Theme.of(context).textTheme;
 
-    return Card(
-      elevation: 1.0,
-      child: Container(
-        decoration: BoxDecoration(
-          color: Colors.white,
-          borderRadius: BorderRadius.all(
-            Radius.circular(5.0),
-          ),
+    return Container(
+      padding: EdgeInsets.all(16.0),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.all(
+          Radius.circular(8.0),
         ),
-        child: TextField(
-          style: textTheme.caption,
-          decoration: InputDecoration(
-            contentPadding: EdgeInsets.all(10.0),
-            border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(5.0),
-              borderSide: BorderSide(
-                color: Colors.white,
-              ),
-            ),
-            enabledBorder: OutlineInputBorder(
-              borderSide: BorderSide(
-                color: Colors.white,
-              ),
-              borderRadius: BorderRadius.circular(5.0),
-            ),
-            hintText: "",
-            prefixIcon: Icon(
-              Icons.search,
-              color: Colors.grey,
-            ),
-            suffixIcon: Icon(
-              Icons.filter_list,
-              color: Colors.grey,
-            ),
-            hintStyle: TextStyle(
-              fontSize: 15.0,
-              color: Colors.grey,
+      ),
+      child: TextField(
+        style: textTheme.caption,
+        decoration: InputDecoration(
+          contentPadding: EdgeInsets.all(8.0),
+          border: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(8.0),
+            borderSide: BorderSide(
+              color: Colors.grey[400],
             ),
           ),
-          maxLines: 1,
-          //controller: searchControl,
+          enabledBorder: OutlineInputBorder(
+            borderSide: BorderSide(
+              color: Colors.grey[400],
+            ),
+            borderRadius: BorderRadius.circular(8.0),
+          ),
+          hintText: "",
+          prefixIcon: Icon(
+            Icons.search,
+            color: Colors.grey[400],
+          ),
+          suffixIcon: Icon(
+            Icons.filter_list,
+            color: Colors.grey[400],
+          ),
+          hintStyle: TextStyle(
+            fontSize: 15.0,
+            color: Colors.grey[400],
+          ),
         ),
+        maxLines: 1,
+        //controller: searchControl,
       ),
     );
   }
